+++ mobile/noarch/example/016_internal_replication.js
@@ -18,7 +18,7 @@ exports.run_tests = function() {
     Ti.API.info("created 20 docs");
     
     // use the database name for the source, not db.internalURL
-    var pull = db_target.replicationFromURL(db_source.name);
+    var pull = db_target.createPullReplication(db_source.name);
     pull.addEventListener('change', function(e) {
       assert(!pull.error, "replication error: "+JSON.stringify(pull.error));
       pullDone = !!(!pull.running && (pull.completed >= pull.total));